requirements.txt
###### Requirements without Version Specifiers ######`
pyautogui
pyperclip
opencv-python
mecab-python3
pymlask
pywin32
SpeechRecognition
pyaudio
openai
